backend/requirements.txt
Flask==2.0.1
Flask-cors
Flask-SQLAlchemy==2.5.1
requests==2.25.1
Werkzeug==2.0.1
flask_login==0.5.0
SQLAlchemy==1.4.22
python-dotenv
pg8000
